Correct data

+++ src/components/transaction/View.jsx
@@ -32,19 +32,26 @@ class View extends Component {
 
 	enableEdit = event => {
         const el = event.target.parentElement.parentElement.firstChild.children
-        let editValues = []
-        for(let i = 0; i < el.length; i++){
-            editValues.push(el[i].innerHTML)
-        }
         const inputs = document.querySelectorAll('#insert-form .form-control')
-        for(let i = 0; i < inputs.length; i++){
-            if(i === inputs.length -1){
-                const date = `${editValues[i].slice(3,5)}/${editValues[i].slice(0,2)}/${editValues[i].slice(6,10)}`
-                inputs[i].valueAsDate = new Date(date)
-            } 
-            else inputs[i].value = editValues[i]
+        let editValues = []
+
+        if(el && inputs) {
+            for(let i = 0; i < el.length; i++){
+                editValues.push(el[i].innerHTML)
+            }
+            for(let i = 0; i < inputs.length; i++){
+                if(i === inputs.length -1){
+                    const date = `${editValues[i].slice(3,5)}/${editValues[i].slice(0,2)}/${editValues[i].slice(6,10)}`
+                    inputs[i].valueAsDate = new Date(date)
+                } 
+                else inputs[i].value = editValues[i]
+            }
+            this.setState({ edit: true })
         }
-        this.setState({ edit: true })
+        else {
+            console.log("Could not access elements in DOM")
+        }
+        
     }
 
     disableEdit = () => {
@@ -56,23 +63,29 @@ class View extends Component {
 		let check = window.confirm("Confirma exclusão?")
 		
 		if(check){
-            console.log(check)
             let el = event.target.parentElement.parentElement.firstChild.children
             let editValues = []
-            for(let i = 0; i < el.length; i++){
-                editValues.push(el[i].innerHTML)
-            }
-            
-            await axios.delete(`${baseApiUrl}/transactions/${editValues[0]}`)
-                .catch(err => console.log(err))
+
+            if (el) {
+                for(let i = 0; i < el.length; i++){
+                    editValues.push(el[i].innerHTML)
+                }
                 
-            this.getData()
+                await axios.delete(`${baseApiUrl}/transactions/${editValues[0]}`)
+                    .catch(err => console.log(err))
+                    
+                this.getData()
+            }
+            else {
+                console.log("Could not find the element in DOM")
+            }
 		}
     }
 
     nextPage = () => {
         let currentIndex = this.state.startIndex
         let newIndex = currentIndex + this.offset
+        
         if(newIndex > this.state.transactions.length){
 
         } else {
